inference.py
import os
import cv2
import numpy as np
import torch
import torchvision
from PIL import Image
import glob
from generate_pairs import Estimator3D

class CFRNet(torch.nn.Module):
    def __init__(self):
        super(CFRNet, self).__init__()
        self.conv1 = torch.nn.Conv2d(3, 64, kernel_size=3, padding=1)
        self.relu = torch.nn.ReLU()
        # Add more layers as per your model

    def forward(self, x):
        x = self.relu(self.conv1(x))
        # Add forward pass logic
        return x

def inference(img_path, save_path, aligner, generator_path, estimator_path, face_model_path, batch_size=4):
    device = torch.device('cpu')
    netG = CFRNet()
    netG.load_state_dict(torch.load(generator_path, map_location=device))
    netG.eval()
    netG = netG.to(device)
    
    estimator3d = Estimator3D(batch_size=batch_size, render_size=224, test=True, estimator_path=estimator_path, face_model_path=face_model_path, cuda_id=0)
    
    img_list = glob.glob(os.path.join(img_path, '*.jpg')) + glob.glob(os.path.join(img_path, '*.png'))
    for i in range(0, len(img_list), batch_size):
        batch_imgs = img_list[i:i+batch_size]
        input_img = estimator3d.load_images(batch_imgs)
        if input_img is None:
            continue
        with torch.no_grad():
            output = netG(input_img)
        for j, img_file in enumerate(batch_imgs):
            save_file = os.path.join(save_path, os.path.basename(img_file))
            save_img = output[j].permute(1, 2, 0).cpu().numpy()
            save_img = (save_img * 255).astype(np.uint8)
            save_img = cv2.cvtColor(save_img, cv2.COLOR_RGB2BGR)
            cv2.imwrite(save_file, save_img)

if __name__ == '__main__':
    import argparse
    parser = argparse.ArgumentParser()
    parser.add_argument('--img_path', type=str, required=True)
    parser.add_argument('--save_path', type=str, required=True)
    parser.add_argument('--aligner', type=str, default='None')
    parser.add_argument('--generator_path', type=str, required=True)
    parser.add_argument('--estimator_path', type=str, required=True)
    parser.add_argument('--face_model_path', type=str, default="mmRegressor/BFM/BFM_model_80.mat")
    parser.add_argument('--batch_size', type=int, default=4)
    args = parser.parse_args()
    inference(args.img_path, args.save_path, args.aligner, args.generator_path, args.estimator_path, args.face_model_path, args.batch_size)