train.py
import os
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader
from torchvision import transforms
import numpy as np
from tensorboardX import SummaryWriter

class CFRNet(nn.Module):
    def __init__(self):
        super(CFRNet, self).__init__()
        self.conv1 = nn.Conv2d(3, 64, kernel_size=3, padding=1)
        self.relu = nn.ReLU()
        # Add more layers as per your model

    def forward(self, x):
        x = self.relu(self.conv1(x))
        # Add forward pass logic
        return x

class TwoScaleDiscriminator(nn.Module):
    def __init__(self):
        super(TwoScaleDiscriminator, self).__init__()
        self.conv1 = nn.Conv2d(3, 64, kernel_size=4, stride=2, padding=1)
        self.relu = nn.LeakyReLU(0.2)
        # Add more layers as per your model

    def forward(self, x):
        x = self.relu(self.conv1(x))
        # Add forward pass logic
        return x

class CFRLogger(object):
    def __init__(self, log_dir):
        self.writer = SummaryWriter(log_dir)
        self.step = 0

    def log_scalar(self, tag, value):
        self.writer.add_scalar(tag, value, self.step)

    def log_images(self, tag, images):
        self.writer.add_images(tag, images, self.step)

    def step_inc(self):
        self.step += 1

def train(args):
    device = torch.device('cpu')
    netG = CFRNet().to(device)
    netD = TwoScaleDiscriminator().to(device)
    
    optimizerG = optim.Adam(netG.parameters(), lr=args.lr, betas=(0.5, 0.999))
    optimizerD = optim.Adam(netD.parameters(), lr=args.lr, betas=(0.5, 0.999))
    
    # Assume dataset is defined elsewhere
    dataset = None  # Replace with your dataset
    dataloader = DataLoader(dataset, batch_size=args.batch_size, shuffle=True)
    
    logger = CFRLogger(args.log_dir)
    
    for epoch in range(args.epochs):
        for i, data in enumerate(dataloader):
            real_imgs = data['image'].to(device)
            # Train Discriminator
            optimizerD.zero_grad()
            real_validity = netD(real_imgs)
            fake_imgs = netG(real_imgs)
            fake_validity = netD(fake_imgs.detach())
            d_loss = torch.mean(fake_validity) - torch.mean(real_validity)
            d_loss.backward()
            optimizerD.step()
            
            # Train Generator
            optimizerG.zero_grad()
            fake_validity = netD(fake_imgs)
            g_loss = -torch.mean(fake_validity)
            g_loss.backward()
            optimizerG.step()
            
            # Logging
            logger.log_scalar('d_loss', d_loss.item())
            logger.log_scalar('g_loss', g_loss.item())
            logger.log_images('real_images', real_imgs)
            logger.log_images('fake_images', fake_imgs)
            logger.step_inc()
            
            if i % 100 == 0:
                print(f"[Epoch {epoch}/{args.epochs}] [Batch {i}/{len(dataloader)}] "
                      f"D loss: {d_loss.item():.4f}, G loss: {g_loss.item():.4f}")
        
        # Save checkpoint
        torch.save(netG.state_dict(), os.path.join(args.save_dir, f'CFRNet_G_ep{epoch}.pth'))
        torch.save(netD.state_dict(), os.path.join(args.save_dir, f'CFRNet_D_ep{epoch}.pth'))

if __name__ == '__main__':
    import argparse
    parser = argparse.ArgumentParser()
    parser.add_argument('--data_dir', type=str, required=True)
    parser.add_argument('--save_dir', type=str, required=True)
    parser.add_argument('--log_dir', type=str, required=True)
    parser.add_argument('--lr', type=float, default=0.0002)
    parser.add_argument('--batch_size', type=int, default=4)
    parser.add_argument('--epochs', type=int, default=100)
    args = parser.parse_args()
    train(args)