requirements.txt
pillow>=7.2.0
opencv-python==4.4.0.44
pytorch3d==0.2.5
scipy==1.5.4
tensorboard==2.4.1